requirements.txt
undetected-chromedriver>=3.5.0
selenium>=4.15.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-telegram-bot>=20.7
python-dotenv>=1.0.0
